Move common packages into bitbucket package and remove println
