+++ src/main/scala/com/codacy/client/bitbucket/client/Request.scala
@@ -1,3 +1,3 @@
-package com.codacy.client.client
+package com.codacy.client.bitbucket.client
 
 case class Request[T](url: String, classType: Class[T])
