+++ src/main/scala/com/codacy/client/bitbucket/client/ResponseError.scala
@@ -1,4 +1,4 @@
-package com.codacy.client.client
+package com.codacy.client.bitbucket.client
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
