+++ src/main/scala/com/codacy/client/bitbucket/util/CommitHelper.scala
@@ -1,4 +1,4 @@
-package com.codacy.client.util
+package com.codacy.client.bitbucket.util
 
 object CommitHelper {
   def anchor(commitUUID: String): String = {
