+++ src/main/scala/com/codacy/client/bitbucket/v1/BuildStatus.scala
@@ -1,6 +1,6 @@
 package com.codacy.client.bitbucket.v1
 
-import com.codacy.client.util.JsonEnumeration
+import com.codacy.client.bitbucket.util.JsonEnumeration
 import play.api.libs.json.Json
 
 object CommitStatus extends JsonEnumeration {
