+++ src/main/scala/com/codacy/client/bitbucket/v1/service/CommitServices.scala
@@ -1,8 +1,8 @@
 package com.codacy.client.bitbucket.v1.service
 
 import com.codacy.client.bitbucket.v1.CommitComment
-import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
-import com.codacy.client.util.CommitHelper
+import com.codacy.client.bitbucket.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.bitbucket.util.CommitHelper
 import play.api.libs.json.{JsNumber, JsObject, JsString}
 
 class CommitServices(client: BitbucketClient) {
