+++ src/main/scala/com/codacy/client/bitbucket/v1/service/RepositoryServices.scala
@@ -1,7 +1,7 @@
 package com.codacy.client.bitbucket.v1.service
 
 import com.codacy.client.bitbucket.v1.{SimpleRepository, SshKey}
-import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.bitbucket.client.{BitbucketClient, Request, RequestResponse}
 import play.api.libs.json.Json
 
 class RepositoryServices(client: BitbucketClient) {
