+++ src/main/scala/com/codacy/client/bitbucket/v1/service/UserServices.scala
@@ -1,6 +1,6 @@
 package com.codacy.client.bitbucket.v1.service
 
-import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.bitbucket.client.{BitbucketClient, Request, RequestResponse}
 import com.codacy.client.bitbucket.v1.{Email, SshKey, User}
 import play.api.libs.json.Json
