+++ src/main/scala/com/codacy/client/bitbucket/v2/service/UrlServices.scala
@@ -1,6 +1,6 @@
 package com.codacy.client.bitbucket.v2.service
 
-import com.codacy.client.client.{BitbucketClient, Request, RequestResponse}
+import com.codacy.client.bitbucket.client.{BitbucketClient, Request, RequestResponse}
 import play.api.libs.json.{JsNull, JsObject}
 
 class UrlServices(client: BitbucketClient) {
